pyproject.toml
[project]
name = "retriever"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "orjson>=3.11.5",
    "pydantic-settings>=2.12.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.30.0",
]

[project.scripts]
tiler = "retriever.components.tiler.cli:run"
victor = "retriever.components.victor.cli:main"
embedder-worker = "retriever.components.embedder_worker.cli:main"
vectordb-service = "retriever.services.vectordb.cli:main"
retriever-service = "retriever.services.retriever.cli:main"
app = "app_streamlit.cli:main"

[project.optional-dependencies]
model = [
    "einops>=0.8.1",
    "ftfy>=6.3.1",
    "pillow>=12.1.0",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
vectordb = [
    "lancedb>=0.26.1",
    "pyarrow>=22.0.0",
]
queue = [
    "pika>=1.3.2",
]
app = [
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pydeck>=0.9.1",
    "streamlit>=1.52.2",
    "pillow>=12.1.0",
]
tiler = [
    "geopandas>=1.1.2",
    "planetary-computer>=1.0.0",
    "pyproj>=3.7.2",
    "pystac-client>=0.9.0",
    "rasterio>=1.4.4",
    "shapely>=2.1.2",
]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.6.0",
]

[tool.uv]
package = true
